requirements.txt
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
numpy
pandas
pybind11
ray
rank-bm25
tensordict<0.6
transformers<4.48
vllm<=0.6.3
wandb
IPython
matplotlib
openai